+++ src/assets/reset.css
@@ -29,6 +29,6 @@ h2 {
   margin-bottom: 5px;
 }
 
-Link {
+.sidebar-link {
   text-decoration: none;
 }
