+++ src/components/Sidebar/Sidebar.css
@@ -1,0 +1,3 @@
+.sidebar-link {
+  text-decoration: none;
+}
