+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { Sidebar } from "./Sidebar/Sidebar";
+export { NoteInput } from "./NoteInput/NoteInput";
+export { SaveNote } from "./SaveNote/SaveNote";
+// export { useNote } from "./Context/noteContext";
