+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export {LandingPage} from "./LandingPage/LandingPage";
+export {Login} from "./Login/Login";
+export {Home} from "./Home/Home";
+export {Archive} from "./Archive/Archive";
